Enforce split between View and Model with Python projects to fix Mono build. (#10748)
* Create main.yml

* Update and rename main.yml to dynamo_mono_build.yml

* Update dynamo_mono_build.yml

* Update dynamo_mono_build.yml

* Update dynamo_mono_build.yml

* remove wpf refs from dsironpython
add new ironPythonProvider to dsiron
put legacy one back in wpf side
add new completion data
PythonNodeModelsWPF will require ironPython at compile time - but not at runtime. API compat.
use shared completion provider that can load different provider cores depedning on engine
dyanamically
dynamo services exposes new interfaces for python internal.
add todo to tests

* complete internal completion core
remove images from dsIronpython - add back to pythonwpf
add additional scope and engine to legacyInterface
refactor IronPythonCompletionData
SharedCompletionProvider class is hooked up.
SharedCompletionProvider is used in script window.

* update mono

* remove avalon edit ref from dsIron

* Update dynamo_mono_build.yml

* Update dynamo_mono_build.yml

* refactor const
add preload to python tests
fix getData

* remove regexs with no use
remove internal regex from legay provider
transfer over summaries
make some methods proper format
add 1 todo
remove todos

* add required ref to dyncore for ilogsource
use ilogsource
cleanup of enum conversion
add summaries
interfaces all internal for now

* add sharedcompletion tests
add new files, split out classes from existing files

* fix case of private method

* remove obsolete methods from entire class that is obsoleted

* address comments

* set a bunch of binaries to copy local false as they were being included in 2 places.

* remove todo

* retire travis job
