+++ src/Libraries/PythonNodeModelsWpf/Properties/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following 
@@ -9,3 +10,5 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCulture("")]
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("5df40db3-627c-43fb-9a0e-0ee422d92485")]
+
+[assembly:InternalsVisibleTo("DynamoPythonTests")]
