+++ src/Libraries/PythonNodeModelsWpf/Resources/class.png
