+++ src/Libraries/PythonNodeModelsWpf/Resources/field.png
