+++ src/Libraries/PythonNodeModelsWpf/Resources/method.png
