+++ src/Libraries/PythonNodeModelsWpf/Resources/namespace.png
