+++ src/Libraries/PythonNodeModelsWpf/Resources/privateMethod.png
