+++ src/Libraries/PythonNodeModelsWpf/Resources/property.png
