+++ src/Libraries/PythonNodeModelsWpf/packages.config
@@ -1,6 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="AvalonEdit" version="4.3.1.9430" targetFramework="net48" />
+  <package id="DynamicLanguageRuntime" version="1.2.2" targetFramework="net48" />
+  <package id="IronPython" version="2.7.9" targetFramework="net48" />
   <package id="FontAwesome.WPF" version="4.7.0.9" targetFramework="net48" />
   <package id="Python.Included" version="3.7.3.4" targetFramework="net48" />
   <package id="Python.Runtime.NETStandard" version="3.7.0" targetFramework="net48" />
